requirements.txt
pymysql
sqlalchemy
pycryptodome
rich
pyperclip